Move QR parts from extendedSession into QrLogin.py
